+++ core-base-fe/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @moonlight-labs/core-base-fe
 
+## 0.2.8
+
+### Patch Changes
+
+- base enhancements
+
 ## 0.2.7
 
 ### Patch Changes