+++ core-notifications-fe/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @moonlight-labs/core-notifications-fe
 
+## 0.2.9
+
+### Patch Changes
+
+- base enhancements
+- Updated dependencies
+  - @moonlight-labs/core-base-fe@0.2.8
+
 ## 0.2.8
 
 ### Patch Changes